requirements.txt
paramiko
requests
python-dotenv
xmltodict
rich